[#1303] Redirect APIv2 requests to `os-api-legacy`
Our backwards-compatible layer has some issues that are breaking a few of our
API users. To give us some time to fix them, this patch redirects calls to any
APIv2 endpoint to our old instance. The endpoints it deals with are:

* /api/2/aggregate
* /api/2/search
* /datasets.json
* /{dataset}.json
* /{dataset}/model.json
* /{dataset}/dimensions.json
* /{dataset}/{dimension}.json
* /{dataset}/{dimension}/{name}.json
* /{dataset}/entries/{id}.json
* /api/mytax
* /api/2/new (Unused, returns HTTP Code 410)

The last endpoint simply returns 410 because our old OpenSpending instance is in
read-only, so only calls that don't modify state are supported.

openspending/openspending#1303